django/dj_app/requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
gunicorn